update to tweak the saving of new radar types

+++ src/main/java/com/pucksandprogramming/technologyradar/domainmodel/UserType.java
@@ -2,7 +2,7 @@ package com.pucksandprogramming.technologyradar.domainmodel;
 
 public class UserType
 {
-    public static int Free = 0;
-    public static int Subscriber_Trial = 1;
-    public static int Subscriber = 2;
+    public static Long Free = 0L;
+    public static Long Subscriber_Trial = 1L;
+    public static Long Subscriber = 2L;
 }
